requirements.txt
torch
openai-whisper
sounddevice
soundfile
numpy
scipy
